Clearing out cruft
