+++ src/components/index.tsx
@@ -3,4 +3,3 @@ export * from "./CredentialEditor";
 export * from "./IssueForm";
 export * from "./VerifyForm";
 export * from "./Props";
-export * from "./Tabs";
