+++ src/pages/index.tsx
@@ -1,4 +1,3 @@
 export * from "./Issue";
-export * from "./VerifiableCredentialEdit";
 export * from "./Verify";
 export * from "./About";
