+++ src/utils/constants.ts
@@ -7,23 +7,3 @@ export const NAV_SIDEBAR_ICONS = [
   { icon: "icon-correct", link: "/verify", label: "Verify", },
   { icon: "icon-qr-code", link: "/about", label: "About", },
 ];
-
-export const MAIN_PAGE_TABS = ["source", "table", "card", "qr code"];
-
-export const BUTTON_BOX_SHADOW = `box-shadow: -10px 10px 20px rgba(58, 31, 134, 0.2),
-10px -10px 20px rgba(58, 31, 134, 0.2),
--10px -10px 20px rgba(86, 47, 200, 0.9),
-10px 10px 25px rgba(58, 31, 134, 0.9),
-inset -10px 10px 20px rgba(58, 31, 134, 0),
-inset 10px -10px 20px rgba(58, 31, 134, 0),
-inset -10px -10px 20px rgba(86, 47, 200, 0),
-inset 10px 10px 25px rgba(58, 31, 134, 0);`;
-
-export const BUTTON_INSET_BOX_SHADOW = `box-shadow: 1px 1px 2px rgba(86, 47, 200, 0.3),
--1px -1px 2px rgba(58, 31, 134, 0.5), 
--10px -10px 20px rgba(86, 47, 200, 0),
-10px 10px 25px rgba(58, 31, 134, 0),
-inset -10px 10px 20px rgba(58, 31, 134, 0.2),
-inset 10px -10px 20px rgba(58, 31, 134, 0.2),
-inset -10px -10px 20px rgba(86, 47, 200, 0.9),
-inset 10px 10px 25px rgba(58, 31, 134, 0.9);`;
